cpp_agent/requirements.txt
openai>=1.30.0
gitpython>=3.1.40
